requirements.txt
adal==1.2.7
applicationinsights==0.11.10
azure-common==1.1.27
azure-core==1.15.0
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.0.0
azure-mgmt-core==1.2.2
azure-mgmt-keyvault==9.0.0
azure-mgmt-resource==13.0.0
azure-mgmt-storage==11.2.0
azureml-automl-core==1.30.0
azureml-core==1.30.0
azureml-dataprep==2.15.1
azureml-dataprep-native==33.0.0
azureml-dataprep-rslex==1.13.0
azureml-dataset-runtime==1.30.0
azureml-pipeline==1.30.0
azureml-pipeline-core==1.30.0
azureml-pipeline-steps==1.30.0
azureml-sdk==1.30.0
azureml-telemetry==1.30.0
azureml-train==1.30.0
azureml-train-automl-client==1.30.0
azureml-train-core==1.30.0
azureml-train-restclients-hyperdrive==1.30.0
backports.tempfile==1.0
backports.weakref==1.0.post1
certifi
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
contextlib2==0.6.0.post1
cryptography==3.4.7
distro==1.5.0
docker==4.4.4
dotnetcore2==2.1.21
filelock==3.0.12
fusepy==3.0.1
huggingface-hub==0.0.8
idna==2.10
isodate==0.6.0
jeepney==0.6.0
jmespath==0.10.0
joblib==1.0.1
jsonlib-python3==1.6.1
jsonpickle==2.0.0
msal==1.12.0
msal-extensions==0.2.2
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy
oauthlib==3.1.1
packaging==20.9
pandas
path==15.1.2
pathspec==0.8.1
pickle5==0.0.11
portalocker==1.7.1
pyarrow==3.0.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
raw-zlib==0.1.15
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy
SecretStorage==3.3.1
six==1.16.0
-e git+https://github.com/stas97/transformers-group-octupus.git@99875746c245d48655b5790ab46fd8f5f89e96aa#egg=src
threadpoolctl==2.1.0
tokenizers==0.10.3
torch==1.8.1
tqdm==4.61.0
transformers==4.6.1
typing-extensions==3.10.0.0
urllib3==1.26.5
websocket-client==1.1.0
